Fix in(x, r::Range{<:Integer}) when crossing types
Previously, failures would happen when checking whether a non-real or
irrational value was in a range, contrary to what happens with an Array.
This was inconsistent with the fallback definition any(y -> y == x, itr).

To fix this, restrict the signature of optimized methods to Real, which are
the only types for which we can be certain the operations are supported.